requirements.txt
aiodns==2.0.0
aiofiles==0.7.0
aiohttp==3.7.4.post0
appdirs==1.4.4
asgiref==3.3.4
astunparse==1.6.3
async-timeout==3.0.1
asyncpg==0.22.0
attrs==20.3.0
beautifulsoup4==4.9.3
better-profanity==0.7.0
black==20.8b1
blinker==1.4
braceexpand==0.1.6
cached-property==1.5.2
cachelib==0.4.1
cchardet==2.1.7
certifi==2020.12.5
cffi==1.14.5
chardet==3.0.4
click==7.1.2
colorama==0.4.4
cycler==0.10.0
discord-ext-ipc==2.0.0
discord-ext-menus @ git+https://github.com/Rapptz/discord-ext-menus@4429b564fd8f031deaccd60caa5be7320e248816
discord.py @ git+https://github.com/Rapptz/discord.py@45d498c1b76deaf3b394d17ccf56112fa691d160
Django==3.2
docopt==0.6.2
future==0.18.2
geographiclib==1.50
geopy==2.1.0
graphviz==0.16
h11==0.12.0
h2==4.1.0
hpack==4.0.0
humanize==3.2.0
hypercorn==0.13.2
hyperframe==6.0.1
idna==2.10
import-expression==1.1.4
itsdangerous==2.0.1
Jinja2==3.0.3
jishaku==1.20.0.220
kiwisolver==1.3.1
MarkupSafe==2.0.1
matplotlib==3.4.1
multidict==4.7.6
mypy-extensions==0.4.3
nest-asyncio==1.5.1
numpy==1.20.2
objgraph==3.5.0
opuslib==3.0.1
pandas==1.3.2
parsedatetime==2.6
pathspec==0.8.1
Pillow==8.2.0
priority==2.0.0
psutil==5.8.0
psycopg2-binary==2.8.6
pycares==3.1.1
pycodestyle==2.7.0
pycparser==2.20
PyNaCl==1.4.0
pyparsing==2.4.7
python-dateutil==2.8.1
pytz==2021.1
quart==0.16.2
Quart-Session==1.0.3
regex==2021.4.4
requests==2.25.1
six==1.15.0
soupsieve==2.2.1
speedtest==0.0.1
spotipy==2.19.0
sqlparse==0.4.1
tdqm==0.0.1
timeago==1.0.15
toml==0.10.2
topggpy==1.3.0
tqdm==4.62.2
typed-ast==1.4.2
typing-extensions==3.7.4.3
tzlocal==2.1
Unidecode==1.2.0
urllib3==1.26.4
Werkzeug==2.0.2
wsproto==1.0.0
yarg==0.1.9
yarl==1.6.3
youtube-dl==2021.6.6
